Pre-processing Script on Local Sys
This shell script starts the submission. I calls POSCAR.py to parse the CIF file into a POSCAR file. It then copies the POSCAR file to the scratch folder on the SSH server. After that it makes a new folder in scratch and moves the POSCAR file into there. Then it calls on the script contained on the SSH server to modify the INCAR-1 and 2 files and then submits it. The only issue is that you can't name your folder with a parenthesis for now.